+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 [unreleased]
+- Enhancement: Firmware updater now support Makera Z1. Detects if fw is bundled LPC+ESP, LPC-only, or ESP-only and updates using the correct method
 - Enhancement: Improve update popup UI and retrieve version data from the GitHub API
 - Enhancement: Replace the nested Remote/Local file popup with a single file browser
 - Enhancement: Auto enclosure/LED lights on connect and off on disconnect or app close. Controller setting is available to enable/disable this feature, default is enabled.